requirements.txt
google-generativeai==0.4.1
pymupdf==1.26.6 
pytesseract==0.3.13
Pillow==11.0.0
protobuf==4.25.3 
requests==2.31.0 
tqdm==4.66.1
urllib3==2.1.0
six==1.16.0